Fix elapsed_ms integer overflow causing main station play recording failure
elapsed_ms was typed as INTEGER (max ~2.1B ms) but plays left open
for extended periods (e.g. stuck play #7602 open since Sep 2025)
produced values exceeding int32 range (~147 days = 12.75B ms).
Every track_change for the main station was failing, so no new plays
were recorded for months. Fix: change column to BIGINT + migration.

+++ services/api/app/models/plays.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey, Text
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, ForeignKey, Text
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -16,7 +16,7 @@ class Play(Base):
     liquidsoap_id = Column(String(100), nullable=True, index=True)  # Liquidsoap track ID
     started_at = Column(DateTime(timezone=True), nullable=False, index=True)
     ended_at = Column(DateTime(timezone=True), nullable=True, index=True)
-    elapsed_ms = Column(Integer, nullable=True)  # How much was actually played
+    elapsed_ms = Column(BigInteger, nullable=True)  # How much was actually played
     
     # Play context
     play_position = Column(Integer, nullable=True)  # Position in playlist/queue
